pyproject.toml
[project]
name = "calculator-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lark-parser>=0.12.0",
    "pytest>=7.0",
    "pyyaml>=6.0",
]
